# openapi.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "Threads-like Application API",
		"description": "This is a Threads-like application backend API that allows users to set a username, post content, follow other users, and interact via comments and likes. No formal authentication required - usernames are stored in browser localStorage.",
		"version": "1.0.0",
		"contact": {
			"name": "API Support",
			"email": "support@example.com"
		}
	},
	"servers": [
		{
			"url": "https://api.example.com/v1",
			"description": "Production server"
		},
		{
			"url": "https://staging-api.example.com/v1",
			"description": "Staging server"
		}
	],
	"paths": {
...
		"/posts/{postId}/comments": {
			"get": {
				"tags": ["Posts"],
				"summary": "Get comments for a post",
				"description": "Retrieve all comments on a specific post",
				"operationId": "get_post_comments_posts__postId__comments_get",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to get comments from",
						"schema": { "type": "integer", "title": "Post ID" }
					},
					{
						"name": "page",
						"in": "query",
						"description": "Page number for pagination (1-indexed)",
						"schema": {
							"type": "integer",
							"default": 1,
							"minimum": 1
						}
					},
					{
						"name": "limit",
						"in": "query",
						"description": "Number of items per page",
						"schema": {
							"type": "integer",
							"default": 10,
							"minimum": 1,
							"maximum": 100
						}
					}
				],
				"responses": {
					"200": {
						"description": "Comments retrieved successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/CommentList" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"post": {
				"tags": ["Posts"],
				"summary": "Add a comment to a post",
				"description": "Create a new comment on a specific post",
				"operationId": "create_comment_posts__postId__comments_post",
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"required": true,
						"description": "ID of the post to comment on",
						"schema": { "type": "integer", "title": "Post ID" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/CommentCreate" }
						}
					},
					"description": "Comment content to add"
				},
				"responses": {
					"201": {
						"description": "Comment added successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Comment" }
							}
						}
					},
					"400": {
						"description": "Bad Request - Invalid input data",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Post does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
		"/comments/{commentId}": {
			"delete": {
				"tags": ["Comments"],
				"summary": "Delete a comment",
				"description": "Remove a comment created by the authenticated user",
				"operationId": "delete_comment_comments__commentId__delete",
				"parameters": [
					{
						"name": "commentId",
						"in": "path",
						"required": true,
						"description": "ID of the comment to delete",
						"schema": { "type": "integer", "title": "Comment ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": true,
						"description": "Username for authentication",
						"schema": { "type": "string" }
					}
				],
				"responses": {
					"200": {
						"description": "Comment deleted successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/SuccessResponse" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Not your comment or invalid session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Comment does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			},
			"put": {
				"tags": ["Comments"],
				"summary": "Update a comment",
				"description": "Modify content of a comment created by the authenticated user",
				"operationId": "update_comment_comments__commentId__put",
				"parameters": [
					{
						"name": "commentId",
						"in": "path",
						"required": true,
						"description": "ID of the comment to update",
						"schema": { "type": "integer", "title": "Comment ID" }
					},
					{
						"name": "username",
						"in": "query",
						"required": true,
						"description": "Username for authentication",
						"schema": { "type": "string" }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/CommentUpdate" }
						}
					},
					"description": "Updated comment content"
				},
				"responses": {
					"200": {
						"description": "Comment updated successfully",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/Comment" }
							}
						}
					},
					"401": {
						"description": "Unauthorized - Not your comment or invalid session",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"404": {
						"description": "Not Found - Comment does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponse" }
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/HTTPValidationError" }
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"CommentCreate": {
				"properties": {
					"content": {
						"type": "string",
						"title": "Content",
						"description": "Text content of the comment",
						"minLength": 1,
						"maxLength": 300
					},
					"username": {
						"type": "string",
						"title": "Username",
						"description": "Username of the comment author (from localStorage)"
					}
				},
				"type": "object",
				"required": ["content", "username"],
				"title": "CommentCreate"
			},
			"CommentUpdate": {
				"properties": {
					"content": {
						"type": "string",
						"title": "Content",
						"description": "Updated text content of the comment",
						"minLength": 1,
						"maxLength": 300
					}
				},
				"type": "object",
				"required": ["content"],
				"title": "CommentUpdate"
			},
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": { "$ref": "#/components/schemas/ValidationError" },
						"type": "array",
						"title": "Detail",
						"description": "List of validation errors"
					}
				},
				"type": "object",
				"title": "HTTPValidationError",
				"description": "Error response for validation failures"
			},
...
			"ValidationError": {
				"properties": {
					"loc": {
						"items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
						"type": "array",
						"title": "Location",
						"description": "Location of the validation error"
					},
					"msg": {
						"type": "string",
						"title": "Message",
						"description": "Validation error message"
					},
					"type": {
						"type": "string",
						"title": "Error Type",
						"description": "Type of validation error"
					}
				},
				"type": "object",
				"required": ["loc", "msg", "type"],
				"title": "ValidationError",
				"description": "Single validation error details"
			},
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string",
						"description": "Error message"
					},
					"status_code": {
						"type": "integer",
						"description": "HTTP status code"
					}
				},
				"required": ["error", "status_code"],
				"title": "ErrorResponse",
				"description": "Standard error response format"
			},
			"SuccessResponse": {
				"type": "object",
				"properties": {
					"message": {
						"type": "string",
						"description": "Success message"
					}
				},
				"required": ["message"],
				"title": "SuccessResponse",
				"description": "Standard success response format"
			},
...
			"Comment": {
				"type": "object",
				"properties": {
					"id": {
						"type": "integer",
						"description": "Comment ID"
					},
					"content": {
						"type": "string",
						"description": "Comment content"
					},
					"post_id": {
						"type": "integer",
						"description": "ID of the post this comment belongs to"
					},
					"author": {
						"type": "object",
						"properties": {
							"id": {
								"type": "integer",
								"description": "Author user ID"
							},
							"username": {
								"type": "string",
								"description": "Author username"
							},
							"profile_image_url": {
								"type": "string",
								"format": "uri",
								"description": "URL to author's profile image"
							}
						},
						"required": ["id", "username"],
						"description": "Comment author information"
					},
					"created_at": {
						"type": "string",
						"format": "date-time",
						"description": "Comment creation timestamp"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time",
						"description": "Last comment update timestamp"
					}
				},
				"required": ["id", "content", "post_id", "author", "created_at"],
				"title": "Comment",
				"description": "Comment information"
			},
			"CommentList": {
				"type": "object",
				"properties": {
					"items": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/Comment"
						},
						"description": "List of comments"
					},
					"total": {
						"type": "integer",
						"description": "Total number of comments"
					},
					"page": {
						"type": "integer",
						"description": "Current page number"
					},
					"size": {
						"type": "integer",
						"description": "Number of items per page"
					},
					"pages": {
						"type": "integer",
						"description": "Total number of pages"
					}
				},
				"required": ["items", "total", "page", "size", "pages"],
				"title": "CommentList",
				"description": "Paginated list of comments"
			},
...
		}
	}
}